+++ src/main/java/kitchenpos/order/ui/OrderRestController.java
@@ -1,7 +1,6 @@
 package kitchenpos.order.ui;
 
-import kitchenpos.order.application.OrderCrudService;
-import kitchenpos.order.application.OrderStatusService;
+import kitchenpos.order.application.OrderService;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.dto.OrderCreateRequest;
 import kitchenpos.order.dto.OrderResponse;
@@ -14,17 +13,15 @@ import java.util.List;
 
 @RestController
 public class OrderRestController {
-    private final OrderCrudService orderCrudService;
-    private final OrderStatusService orderStatusService;
+    private final OrderService orderService;
 
-    public OrderRestController(final OrderCrudService orderCrudService, final OrderStatusService orderStatusService) {
-        this.orderCrudService = orderCrudService;
-        this.orderStatusService = orderStatusService;
+    public OrderRestController(final OrderService orderService) {
+        this.orderService = orderService;
     }
 
     @PostMapping("/api/orders")
     public ResponseEntity<OrderResponse> create(@RequestBody final OrderCreateRequest request) {
-        final OrderResponse created = orderCrudService.create(request);
+        final OrderResponse created = orderService.create(request);
         final URI uri = URI.create("/api/orders/" + created.getId());
         return ResponseEntity.created(uri)
                 .body(created);
@@ -33,11 +30,11 @@ public class OrderRestController {
     @GetMapping("/api/orders")
     public ResponseEntity<List<Order>> list() {
         return ResponseEntity.ok()
-                .body(orderCrudService.list());
+                .body(orderService.list());
     }
 
     @PutMapping("/api/orders/{orderId}/order-status")
     public ResponseEntity<Order> changeOrderStatus(@PathVariable final Long orderId, @RequestBody final OrderStatusChangeRequest request) {
-        return ResponseEntity.ok(orderStatusService.changeOrderStatus(orderId, request));
+        return ResponseEntity.ok(orderService.changeOrderStatus(orderId, request));
     }
 }
